+++ plugin-dev/Source/Sentry/Public/SentrySettings.h
@@ -43,6 +43,17 @@ enum class ESentryDatabaseLocation : uint8
 	ProjectUserDirectory
 };
 
+UENUM(BlueprintType)
+enum class ESentryAndroidCrashBackend : uint8
+{
+	// Capture crashes using the sentry-native NDK signal handler only
+	NdkSignalHandler,
+	// Capture crashes using Android tombstones only
+	TombstoneOnly,
+	// Capture crashes using NDK integration with additional context merged from tombstones (recommended)
+	TombstoneMergedWithNdk,
+};
+
 USTRUCT(BlueprintType)
 struct FAutomaticBreadcrumbs
 {
@@ -358,6 +369,10 @@ class SENTRY_API USentrySettings : public UObject
 	UPROPERTY(Config, EditAnywhere, BlueprintReadWrite, Category = "General|Mobile",
 		Meta = (DisplayName = "Enable ANR error tracking", Tooltip = "Flag indicating whether to enable tracking of ANR (app not responding) errors."))
 	bool EnableAppNotRespondingTracking;
+
+	UPROPERTY(Config, EditAnywhere, Category = "General|Mobile",
+		Meta = (DisplayName = "Android crash capturing backend", ToolTip = "Mechanism used to capture Android crashes and enrich them with additional context."))
+	ESentryAndroidCrashBackend AndroidCrashBackend;
 
 	UPROPERTY(Config, EditAnywhere, Category = "General|Performance Monitoring",
 		Meta = (DisplayName = "Enable tracing", ToolTip = "Flag indicating whether to enable tracing for performance monitoring."))